+++ src/V2/Tests/TestFiles/EntityFrameworkCore/ExternalInterfaceTestsAsyncEfc.cs
@@ -25,7 +25,7 @@ namespace ServiceQuery.Xunit
         }
 
         [Fact]
-        public async Task ExternalInterfaceStandardTest()
+        public void ExternalInterfaceStandardTest()
         {
             var sourceQueryable = GetTestList();
 
@@ -115,7 +115,7 @@ namespace ServiceQuery.Xunit
         }
 
         [Fact]
-        public async Task NullExternalInterfaceStandardTest()
+        public void NullExternalInterfaceStandardTest()
         {
             var sourceQueryable = GetTestList();
 
@@ -205,7 +205,7 @@ namespace ServiceQuery.Xunit
         }
 
         [Fact]
-        public async Task ExternalInterfaceRequestTest()
+        public void ExternalInterfaceRequestTest()
         {
             var sourceQueryable = GetTestList();
 
@@ -295,7 +295,7 @@ namespace ServiceQuery.Xunit
         }
 
         [Fact]
-        public async Task NullExternalInterfaceRequestTest()
+        public void NullExternalInterfaceRequestTest()
         {
             var sourceQueryable = GetTestList();
 
@@ -385,7 +385,7 @@ namespace ServiceQuery.Xunit
         }
 
         [Fact]
-        public async Task ExternalInterfaceRequestExecuteTest()
+        public void ExternalInterfaceRequestExecuteTest()
         {
             var sourceQueryable = GetTestList();