+++ src/V2/Tests/TestFiles/EntityFrameworkCore/NegativeTestsAsyncEfc.cs
@@ -34,7 +34,7 @@ namespace ServiceQuery.Xunit
     public abstract class NegativeTestsAsyncEfc<T> : BaseTest<T> where T : class, ITestClass, new()
     {
         [Fact]
-        public async Task EndBeforeBegin()
+        public void EndBeforeBegin()
         {
             var sourceQueryable = GetTestList();
             var sq = ServiceQueryRequestBuilder.New()
@@ -59,7 +59,7 @@ namespace ServiceQuery.Xunit
         }
 
         [Fact]
-        public async Task MissingEnd()
+        public void MissingEnd()
         {
             var sourceQueryable = GetTestList();
             var sq = ServiceQueryRequestBuilder.New()
@@ -73,7 +73,7 @@ namespace ServiceQuery.Xunit
         }
 
         [Fact]
-        public async Task MultipleAndOr()
+        public void MultipleAndOr()
         {
             var sourceQueryable = GetTestList();
             var sq = ServiceQueryRequestBuilder.New()
@@ -110,7 +110,7 @@ namespace ServiceQuery.Xunit
         }
 
         [Fact]
-        public async Task PropertyNotFound()
+        public void PropertyNotFound()
         {
             var sourceQueryable = GetTestList();
             var sq = ServiceQueryRequestBuilder.New()
@@ -143,7 +143,7 @@ namespace ServiceQuery.Xunit
         }
 
         [Fact]
-        public async Task PageSizeNotANumber()
+        public void PageSizeNotANumber()
         {
             var sourceQueryable = GetTestList();
             //IQueryable<T> testQueryable;
@@ -209,7 +209,7 @@ namespace ServiceQuery.Xunit
         }
 
         [Fact]
-        public async Task RequestBeginEnd()
+        public void RequestBeginEnd()
         {
             var sourceQueryable = GetTestList();
             IQueryable<T> testQueryable;
@@ -227,7 +227,7 @@ namespace ServiceQuery.Xunit
         }
 
         [Fact]
-        public async Task IncludeCountTwice()
+        public void IncludeCountTwice()
         {
             var sourceQueryable = GetTestList();
             IQueryable<T> testQueryable;
@@ -252,7 +252,7 @@ namespace ServiceQuery.Xunit
         }
 
         [Fact]
-        public async Task IsAggregateNullCheck()
+        public void IsAggregateNullCheck()
         {
             IServiceQueryRequest serviceQuery;
 
@@ -264,7 +264,7 @@ namespace ServiceQuery.Xunit
         }
 
         [Fact]
-        public async Task FilterTypeNotDefined()
+        public void FilterTypeNotDefined()
         {
             IServiceQueryRequest serviceQuery;
 
@@ -281,7 +281,7 @@ namespace ServiceQuery.Xunit
         }
 
         [Fact]
-        public async Task MissingProperty()
+        public void MissingProperty()
         {
             var sourceQueryable = GetTestList();
             IServiceQueryRequest serviceQuery;
@@ -714,7 +714,7 @@ namespace ServiceQuery.Xunit
         }
 
         [Fact]
-        public async Task ExecuteNull()
+        public void ExecuteNull()
         {
             IServiceQueryRequest serviceQuery;
 
@@ -725,7 +725,7 @@ namespace ServiceQuery.Xunit
         }
 
         [Fact]
-        public async Task SyncNegativeIsEqualStandardTest()
+        public void SyncNegativeIsEqualStandardTest()
         {
             var sourceQueryable = GetTestList();
             IQueryable<T> testQueryable;
@@ -919,7 +919,7 @@ namespace ServiceQuery.Xunit
         }
 
         [Fact]
-        public async Task SyncIsEqualRequestTest()
+        public void SyncIsEqualRequestTest()
         {
             var sourceQueryable = GetTestList();
             IQueryable<T> testQueryable;
@@ -1113,7 +1113,7 @@ namespace ServiceQuery.Xunit
         }
 
         [Fact]
-        public async Task SyncNullIsEqualStandardTest()
+        public void SyncNullIsEqualStandardTest()
         {
             var sourceQueryable = GetTestList();
             IQueryable<T> testQueryable;
@@ -1307,7 +1307,7 @@ namespace ServiceQuery.Xunit
         }
 
         [Fact]
-        public async Task SyncNullIsEqualRequestTest()
+        public void SyncNullIsEqualRequestTest()
         {
             var sourceQueryable = GetTestList();
             IQueryable<T> testQueryable;