+++ src/V2/Tests/TestFiles/MongoDb/PagingTestsAsyncMongoDb.cs
@@ -6,7 +6,7 @@ namespace ServiceQuery.Xunit
     public class PagingTestsAsyncMongoDb : BaseTest
     {
         [Fact]
-        public async Task NegativePagingRequestTest()
+        public void NegativePagingRequestTest()
         {
             var sourceQueryable = GetTestList();
             IServiceQueryRequest serviceQuery;
@@ -42,7 +42,7 @@ namespace ServiceQuery.Xunit
         }
 
         [Fact]
-        public async Task PagingStandardTest()
+        public void PagingStandardTest()
         {
             var sourceQueryable = GetTestList();
             IServiceQuery serviceQuery;
@@ -184,7 +184,7 @@ namespace ServiceQuery.Xunit
         }
 
         [Fact]
-        public async Task PagingRequestTest()
+        public void PagingRequestTest()
         {
             var sourceQueryable = GetTestList();
             IServiceQueryRequest serviceQuery;